Refactoring: moved player suppliers closer to their packages
