+++ src/main/java/com/leokom/chess/player/legal/brain/simple/SimplePlayerSupplier.java
@@ -1,4 +1,4 @@
-package com.leokom.chess;
+package com.leokom.chess.player.legal.brain.simple;
 
 import com.leokom.chess.player.Player;
 import com.leokom.chess.player.legal.LegalPlayer;
