+++ src/main/java/com/leokom/chess/player/winboard/WinboardPlayerSupplier.java
@@ -1,11 +1,11 @@
-package com.leokom.chess;
+package com.leokom.chess.player.winboard;
 
 import com.leokom.chess.player.Player;
 import com.leokom.chess.player.winboard.WinboardPlayer;
 
 import java.util.function.Supplier;
 
-class WinboardPlayerSupplier implements Supplier<Player> {
+public class WinboardPlayerSupplier implements Supplier<Player> {
 
     @Override
     public Player get() {
